(PC-35064)[PRO] feat: remove WIP_ENABLE_OFFER_ADDRESS_COLLECTIVE from e2e tests
